core(stat): move implementations into .hpp file w/o changes
